requirements.txt
# Claudine Server v1 - Python Dependencies
# Clean Architecture rebuild with FastAPI

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic[email]==2.5.3
python-dotenv==1.0.0
email-validator==2.1.0

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# HTTP Client
httpx==0.26.0

# Utilities
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0  # For TestClient

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
